Add default scope to hide scratch projects
We want to be able to add projects with `project_type` set to "scratch"
to use in the `experience-cs` Scratch editor, but we don't want to break
any of the existing Code Editor functionality which almost certainly
relies on projects only having `project_type` set to "python" or "html".

We considered changing the scope just in the `ProjectLoader` class, but
it seems as if there are other places in the code which load projects,
so we went with a default scope instead.

Ideally there would already have been a validation or a database
constraint on `Project#project_type` to prevent it being set to an
unknown type. We considered adding it before making this change.
However, we'd need to check the existing data to make sure it didn't
become invalid. So we're going to leave that for now.

It might also be worth considering using an enum for this attribute at
some point.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>