Add scratch-only option for project show endpoint
We want to be able to load Scratch projects from the Experience CS app
without breaking anything in the Code Editor app. This is now possible
by setting `project_type=scratch` in the query string for the
`Api::Projects#show` action.

We've bumped the maxium number of nested rspec contexts up from 4 -> 5
which we don't think is too terrible. It would be quite hard to re-work
the changes in this spec to use less nesting without significant
restructuring which we want to avoid at the moment.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>