Extract string literals -> Project::Types constants
This introduces the following constants:

* `Project::Types::PYTHON`
* `Project::Types::HTML`
* `Project::Types::SCRATCH`

We did contemplate doing this for just the Scratch project type that
we've introduced in this PR, but we thought it would leave the code in a
more confusing state.

I've left the projects YAML in the `lib/tasks/project_components`
sub-directories unchanged. However, it _might_ be worth changing them to
be processed by ERB so that the new constants can be used in them.

I've also left the 'html' string literals representing
`Component#extension` and `Component#lang` unchanged, because these are
different concepts and so should be modelled separately.