Create Experience CS example project on release
In #527, the `projects:create_all` rake task invocation was removed from
the Heroku release, because of concerns it might reset/overwrite
existing projects in the database.

Now that we have a separate & idempotent
`projects:create_experience_cs_example` rake task with much more limited
scope, we're confident that it's safe to invoke this new rake task as
part of the Heroku release.