Extract ProjectScoping concern & Current.project_scope
To remove duplication and simplify the code. This means there is now
only one place where we use the `project_type` param to override the
default scope on Project to give access to scratch projects. What's
more, it should be easier to add this to other controller actions in the
future.

Note that the underlying value of `Current.project_scope` will get reset
to `nil` after every request [1]. Also note that the default value
functionality for these attributes wasn't added until Rails v7.2 [2], so
we have to override the reader method to provide the default value if
the underlying value is `nil`.

[1]: https://api.rubyonrails.org/v7.1.3.4/classes/ActiveSupport/CurrentAttributes.html
[2]: https://github.com/rails/rails/blob/v7.2.0/activesupport/CHANGELOG.md