Add another example Scratch project for Experience CS
We're planning to use this to replace the existing project with
identifier `experience-cs-example`. However, rather than worrying about
renaming the existing one, we're just going to add this new one for now
and change the `experience-cs` seed data to use this new identifier. We
can worry about deleting the original project later.

Co-authored-by: Seb Jacobs <me@sebjacobs.com>