Add invite spec example to diagnose flakey example
The example just before this one keeps failing in the CircleCI
build [1], but I haven't been able to replicate it locally. It's hard
to work out what's going on, because `SchoolTeacher::Invite#call`
swallows any exception. I'm hoping that this new assertion _might_ fail
for the same reason, but then display the exception message to give us
more clues.

[1]: https://app.circleci.com/pipelines/github/RaspberryPiFoundation/editor-api/2359/workflows/ecf1c252-3f22-45e6-b0fe-08363f66c8d9/jobs/4447