Add Api::PublicProjectsController#create
We want to allow experience-cs admins to create "public" projects via
experience-cs administrate UI [1]. When I say "public" projects, I mean
those that are accessible to all users, because `Project#user_id` is set
to `nil`.

Public projects for Python & HTML are currently added via the
`GithubWebhooksController#github_push` [2] action and/or via the
(possibly defunct) `projects:create_all` rake task [3] both of which
make use of `ProjectImporter` [4]. However, these mechanisms are not
suitable for our use case.

This commit introduces a new `Api::PublicProjectsController` which is
somewhat based on `Api::ProjectsController`. However, it differs in a
couple of key respects: only experience-cs admin users should be
authorized to use it; it should not be possible to set
`Project#user_id`; and we want some sensible validation to restrict the
values entered into the administrate UI and provide sensible error
messages to the user. Conceivably this could be achieved by enhancing
the existing `Api::ProjectsController`, but this is already pretty
complicated and I wanted to minimize the chances of breaking existing
behaviour.

I've tried to roughly follow the existing conventions however much I
disagree with them, e.g. I've introduced a `PublicProject::Create` class
in `lib/concepts/public_project/operations/create.rb` along the same
lines as `Project::Create` in `lib/concepts/project/operations/create.rb`.

This is just a skeletal starting point - I plan to flesh it out in
subsequent commits.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/issues/405
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/f348dfc92fdd8b19fbfd5434e7751340f33b4093/app/controllers/github_webhooks_controller.rb#L6-L8
[3]: https://github.com/RaspberryPiFoundation/editor-api/blob/f348dfc92fdd8b19fbfd5434e7751340f33b4093/lib/tasks/projects.rake#L4-L7
[4]: https://github.com/RaspberryPiFoundation/editor-api/blob/f348dfc92fdd8b19fbfd5434e7751340f33b4093/lib/project_importer.rb