Allow skipping of identifier generation
When creating a public project we want the admin user to be able to
specify the identifier and apply validation rules against it. However,
prior to this change, if the supplied identifier was `nil`, the
`before_validation` callback, `Project#generate_identifier`, would kick
in and generate an identifier. Thus the presence validation on
`Project#identifier` never failed.

The new `Project#skip_identifier_generation` attribute gives us the
ability in specific contexts to prevent the identifier being generated
and thus allow the presence validation to fail.