Forbid creation of non-Scratch public projects
Accepting the `product_type` param means that this controller can
legitimately be named more generically, i.e. `PublicProjectsController`
vs e.g. `ScratchPublicProjectsController`. However, the immediate
requirement is for admin users in the experience-cs app to be able to
create public Scratch projects and so I've only implemented the
functionality necessary to support that use case.

If we want to expand the capability of this controller to support the
creation of HTML and/or Python projects, we'll need to support other
params, e.g. `components`.