Introduce PublicProject model
I want to add some validations to the `Project` model to provide some
sensible constraints on the values that an experience-cs admin user
enters into the administrate UI so that I can provide sensible error
messages.

However, since I don't have access [1] to staging or production data for
`editor-api`, I can't be sure that some existing projects would become
invalid. So as a workaround I've implemented this `PublicProject`
wrapper class which allows me to add validations for use only in the
context of the `Api::PublicProjectsController`.

To start with, I've just added a validation to check the format of
`Project#identifier`, although this isn't yet being used in the
controller action.

Hopefully implementing these validations as ActiveModel validations will
make it easier to move them onto the `Project` model once it's deemed
safe to do so.

[1]: https://raspberrypifoundation.slack.com/archives/C08831T17H6/p1747900618263229