Ensure Project#user_id is not set on public project
A public project, i.e. one that is accessible to all users, should not
have `Project#user_id` set. This is already implicitly enforced by
`Api::PublicProjectsController#project_params` which excludes the
`user_id` param. However, I thought it was worth adding an example to
guarantee it remains the case.