Use ProjectLoader in Api::PublicProjectsController
This is closely based on how `Api::ProjectsController` works although
it's not obvious that its behaviour is tested. However, there is a spec
for `ProjectLoader` itself, so I've just added examples to make sure
it's wired up to the controller correctly.

One thing I've added is to raise `ActiveRecord::RecordNotFound` if
`ProjectLoader#load` returns `nil` which seems possible. This will
result in a 404 Not Found response which is what I was already asserting
in `spec/features/public_project/updating_a_public_project_spec.rb`. As
far as I can tell, `Api::ProjectsController` actions which make use of
`ProjectLoader#load` (e.g. the `update` action) do not check for the
project being `nil` and so will probably respond with a 500 Internal
Server Error when the code (e.g. in `Project::Update`) calls a method on
the `nil` project. This doesn't seem great!