Public project update cannot change certain attributes
There's currently no need to be able to *update* the `locale` or
`project_type` from experience-cs, but since we are supplying them for
the `create` action, I think it's useful to highlight that they're not
accepted by the `update` action.

It doesn't make much sense to allow the `user_id` to be updated via this
controller that specifically relates to "public" projects, so I think
it's useful to add an example to highlight that `user_id` is not
accepted by the `update` action.

This behaviour is currently effectively implemented by omitting these
attibute keys from the permitted params in
`Api::PublicProjectsController#update_params`.