Add Api::PublicProjectsController#destroy
We want to allow experience-cs admins to update "public" projects via
experience-cs administrate UI [1]. When I say "public" projects, I mean
those that are accessible to all users, because `Project#user_id` is set
to `nil`.

Public projects for Python & HTML are currently updated via the
`GithubWebhooksController#github_push` [2] action and/or via the
(possibly defunct) `projects:create_all` rake task [3] both of which
make use of `ProjectImporter` [4]. However, it doesn't look as if
there's a way to delete an existing project through that mechanism and
in any case it's not very suitable for our use case.

This commit introduces a new `Api::PublicProjectsController#destroy`
action which is somewhat based on `Api::ProjectsController#destroy`.

As before I've tried to roughly follow the existing conventions however
much I disagree with them.

Unlike in `Api::ProjectsController#destroy`, the action responds with
422 Unprocessable Entity if `Project#destroy` returns `false`.

This is just a skeletal starting point - I plan to flesh it out in
subsequent commits.

However, I have already re-used the `ProjectLoader` for finding the
project by identifier and locale.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/issues/405
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/f348dfc92fdd8b19fbfd5434e7751340f33b4093/app/controllers/github_webhooks_controller.rb#L6-L8
[3]: https://github.com/RaspberryPiFoundation/editor-api/blob/f348dfc92fdd8b19fbfd5434e7751340f33b4093/lib/tasks/projects.rake#L4-L7
[4]: https://github.com/RaspberryPiFoundation/editor-api/blob/f348dfc92fdd8b19fbfd5434e7751340f33b4093/lib/project_importer.rb