Make destroying public projects specs more realistic
The immediate use case for `Api::PublicProjectsController#destroy` is
for experience-cs admin users to destroy a public project.  Since
experience-cs only deals with Scratch projects, it makes sense to use
projects of this type in the specs. However, since we're using
`ProjectLoader` to find the project and the default scope is still in
place to hide Scratch projects, we need to make use of the
`ProjectScoping` concern to use the `project_type` param to set
`Current.project_scope` to include Scratch projects.