Prevent destruction of public project with remixes
The only use case for `Api::PublicProjectsController#destroy` is
currently to allow experience-cs admins to delete a project. Although
the `Project` model allows such a deletion, it's not obvious that it's
entirely sensible. And in any case, while the Learning Team is using
the experience-cs admin UI to create projects for the curriculum,
there's little danger that they will have remixes until they're
associated with lessons, etc. So this seems like a safer option for now.
We can always relax it later if it becomes a problem.