Pass current_user to Project::Create.call
I want to make some new behaviour conditional on the type of user. This
change will make that easier.

Note that I've used the `Identifiable#current_user` method (included
into `ApiController`) to access the user unlike _some_ places in the
`Api::ProjectsController` which use the `@current_user` instance
variable directly. The latter feels dangerous, because the instance
variable is only defined after the first time in a request that
`Identifiable#current_user` is called and the value is memoized. Thus
the places that are using the instance variable directly are relying on
an earlier call to `Identifiable#current_user` (probably in
`load_and_authorize_resource`) which feels risky and confusing. This
seems like a fairly widespread problem across the API controllers, so
I'm not going to attempt to fix it here.