Simplify Project::Create.build_project
There's no need to explicitly set `Project#identifier` here as well as
in the `generate_identifier` `before_validation` in `Project`. While I'm
not very convinced about the use of the `before_validation` callback,
removing that would be a bigger and riskier piece of work. So instead
I'm removing it here safe and relying on the `before_validation` to do
its thing.

Note that I added feature spec coverage to check the project is assigned
a generated identifier in a previous commit.