Allow ExCS admin to create starter projects
This gives users with the "experience-cs-admin" role permission to
create starter (or "public") projects (i.e. projects with no `user_id`
set). I've added examples to the "Creating a project" feature spec to
check this works as intended.

Note that I've had to add support for `User#roles` to
`UserProfileMock#user_to_hash` for the new examples that I've added to
the feature spec.