Fix updating project feature spec
Previously this spec was not representative of how the API was being
used in practice and it was only really working by accident. The PUT
requests in the examples were using the `Project#id` rather than the
`Project#identifier` as they should have been.

The spec was working even though the
`Api::ProjectsController#load_project` before action was *not* finding
the project, because the `load_and_authorize_resource` before action was
then finding the project using `Project.find(params[:id])` and setting
the `@project` instance variable used by the rest of the logic in the
controller.

However, clients of editor-api like editor-standalone use the project
identifier as the resource ID in the URL path [1], so this spec was
completely unrepresentative of how the endpoint was really being used.

In this commit I've changed the examples to use the `Project#identifier`
as the resource ID in the PUT requests. This means that the
`Api::ProjectsController#load_project` before action now finds the
project by identifier and since it sets the `@project` instance variable
the `load_and_authorize_resource` before action no longer attempts to
load the resource, it just does the authorize step using the already
loaded project.

In order to make this work reliably, I had to explicitly set the
`Project#locale` to one of the fallback locales in `ProjectLoader` [2],
i.e. 'en' or `nil`. Otherwise, the random locale selected in the project
factory [3] meant that sometimes the `load_project` before action (which
uses the `ProjectLoader` did not find the project and the
`load_and_authorize_resource` before action raised an
`ActiveRecord::RecordNotFound` resulting in a 404 Not Found.

Now the spec is no longer making use of the randome locale from the
factory, I've taken the opportunity to add some new examples
demonstrating the behaviour when the project has different locales. This
effectively demonstrates that `load_project` is wired up to
`ProjectLoader` correctly.

As an aside, I'm not convinced that having the factory select a locale
at random is a good idea. I found it very confusing when it led to
undeterministic specs. However, I'll leave that for another time.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/blob/1d4375635cb6890794732072d608dbd4b05b3bb0/src/utils/apiCallHandler/projects.js#L16
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/b4bd337d09a88b1f41ecdc13136f7d11da0dcf89/lib/project_loader.rb#L8
[3]: https://github.com/RaspberryPiFoundation/editor-api/blob/b4bd337d09a88b1f41ecdc13136f7d11da0dcf89/spec/factories/project.rb#L9