Allow ExCS admin to destroy starter projects
This gives users with the "experience-cs-admin" role permission to
update starter (or "public") projects (i.e. projects with no `user_id`
set). I've added an example to the destroy project request spec to check
this works as intended.

There are obviously some dangers associated with destroying a starter
project, but I think it's still useful for an ExCS admin to be able to
do it.