Allow anonymous access to Api::ProjectsController#index
Previously, while it was possible for an anonymous user to view a single
public project (i.e. where `Project#user_id` is `nil`) via
`Api::ProjectsController#show`, it was *not* possible for an anonymous
user to list all public projects.

This removes the `authorize_user` before action for the
`Api::ProjectsController#index` action and changes the permissions for
an anonymous user to allow `read` access (equivalent to both `show` &
`index` access) instead of just `show` access to public projects.

I can't see a problem with exposing projects in this way given that they
are individually exposed and I couldn't see any explanation in the git
history to explain why this wasn't previously allowed.

My motivation for making the changes is that on the projects page in the
experience-cs admin UI we want to only list public projects, because
this is where the Learning Team create starter Scratch projects for the
curriculum. And while we could theoretically cache the `Project#user_id`
in the experience-cs database, we then have to worry about when to
invalidate that cache. I'd far rather we avoid that extra complexity if
we can.