Add examples to fix pending example warning
The `SchoolProject` was added in #490, but for some reason this spec was
left with only a pending example. The latter is displayed in a warning
list when you run the specs and constantlt makes me think something is
wrong. To avoid this, I've implemented a couple of examples to test the
most obvious behaviour of the `SchoolProject` model which seems better
than nothing.

These examples shouldn't be taken as an endorsement of the existing
behaviour of or the desired behaviour of `SchoolProject`. For example, I
wonder whether it should have some validations.