Configure CEfE generator randomness with RSpec seed
We still have at least one flakey spec [1] and I noticed it depends on
the `verified_school` factory which uses the
`ForEducationCodeGenerator.generate` method to set the `School#code`. On
the offchance this is causing some kind of non-deterministic problem and
to make that generator more deterministic in specs, I thought it would
be worthwhile using the RSpec seed to seed the random number generator
(RNG) used by `ForEducationCodeGenerator.generate`.

The changes in this commit allow the RNG to be overridden and adds file
in `spec/support` to do this in specs so it uses the RSpec seed.

[1]: https://github.com/RaspberryPiFoundation/editor-api/blob/f397e870f2a33cce1f53b9104c52314f5233572c/spec/concepts/school_teacher/invite_spec.rb#L14-L17